Rename complete profile file
